requirements.txt
Eve
py-bcrypt
